rename package dataTest to data rename method isPageOpen to isPageOpened rename method Setup to setUp change value of MAX_RETRY add retryAnalyzer to test accountShouldBeCreated
