+++ src/test/java/pages/BasePage.java
@@ -12,5 +12,5 @@ public abstract class BasePage {
         wait = new WebDriverWait(driver, 20);
     }
 
-    public abstract BasePage isPageOpen();
+    public abstract BasePage isPageOpened();
 }
